Stringify seed when creating DOM
When `karma.conf.js` is configured to use a number for a string, the reporter will crash at the end, causing test runs and coverage information to fail even if all tests pass. This is reported in #19. It's also a fairly common error: [Error: Failed to execute 'appendChild' on 'Node': parameter 1 is not of type 'Node'](https://stackoverflow.com/questions/27079598/error-failed-to-execute-appendchild-on-node-parameter-1-is-not-of-type-no) has 334k views on Stack Overflow.

This fixes that error by converting the seed to a string when it's passed to createDom. To be perhaps far more safe than necessary, we avoid assuming that the seed value has a `toString()` method available and instead use [the global `String()` function](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/String/String) (not constructor).